triggerdata: change "Category" into an array
Category was a single-valued dictionary before.
The game actually only has a single value there.
My intention was to force your hand down the line to accept multiple
category entries, so custom categories can be used at some point.

Arrays is the way to go, not dictionaries with made up key names to be
iterated through.